Update
Code cleanup for compilation with Clang and MSVC
Add user service layer
Add an option to disable default key behavior
Disable notifier surface from rendering
Instanciable Tracer
Refactor Tracer initilization with logger to use std location
Update primary services identification
Improve Window fullscreen management
Adds C++ constants instead of macros
Add the visibility check per surfaces for input in overlay manager.
Enable Wayland support.
Update some constexpr usage.
GTK3 removed.
Add dialog box cancelable state.
Crash fixed on some situation.

+++ src/Graphics/Material/StandardResource.hpp
@@ -195,7 +195,7 @@ namespace EmEn::Graphics::Material
 
 			/** @copydoc EmEn::Graphics::Material::Interface::frameIndexAt() */
 			[[nodiscard]]
-			size_t frameIndexAt (uint32_t sceneTime) const noexcept override;
+			uint32_t frameIndexAt (uint32_t sceneTime) const noexcept override;
 
 			/** @copydoc EmEn::Graphics::Material::Interface::enableBlending() */
 			void enableBlending (BlendingMode mode) noexcept override;
